requirements.txt
anyio==3.7.1
asgiref==3.7.2
autopep8==2.0.2
certifi==2023.5.7
Django==4.2.3
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.4.0
exceptiongroup==1.1.2
h11==0.14.0
httpcore==0.17.3
idna==3.4
Pillow==10.0.0
psycopg2==2.9.5
pycodestyle==2.10.0
PyJWT==2.7.0
pytz==2022.7.1
sniffio==1.3.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2023.3
coreapi==2.3.3